+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# v4.3.4 @ 2024-10-22
+
+### ✨ Features
+
+* 优化：`参数构建器` 的 `field(..)` 方法，兼容直接使用 **返回类型不一致** 的三元表达式直接入参。
+
+
+```java
+var params = MapUtils.builder()
+        // ifTrue 若真，返回 List, 否则返回 int，类型一致，v4.3.4 开始兼容这种写法
+        .field(User::getId, ifTrue ? List.of(1,2,3) : 4)
+        .build();
+```
+
+```java
+var params = MapUtils.builder()
+        // ifTrue 若真，返回 int[], 否则返回 int，类型一致，v4.3.4 开始兼容这种写法
+        .field(User::getId, ifTrue ? new int[] {1,2,3} : 4)
+        .build();
+```
+
+```java
+var params = MapUtils.builder()
+        // ifTrue 若真，返回 Integer[], 否则返回 int，类型一致，v4.3.4 开始兼容这种写法
+        .field(User::getId, ifTrue ? new Integer[] {1,2,3} : 4)
+        .build();
+```
+
 # v4.3.3 @ 2024-10-08
 
 ### ✨ Features